project: Initial commit
This was made possible due to:
    docs.flatpak.org
    GloriousEggroll
    johnramsden
    TingPing

It features the RDNN notation for the game (but i'm not sure if it's
correct to use the "_" to delegate the spaces), and a proper sandbox,
which only has access to $XDG_DOCUMENTS_DIR (which is the only user
folder) the games access.

It also features the staging patches, and pba patches with knobs and
switches (thanks acomminos and FireRat), which greatly enhances the
in-game fps.

Hope it serves you well.

Signed-off-by: Rafael Julio <rafael_julio@protonmail.com>